[BpkChatbotInput] Inline default send button and remove SendButton directory
The SendButton wrapper had no remaining logic after the composable
refactor, so inline it directly in BpkChatbotInputInput.
